backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
streamlit==1.44.1
python-dotenv==1.1.0
httpx==0.28.1
requests==2.32.3
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.63
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.8
google-generativeai==0.8.5
pinecone==7.0.2
PyPDF2==3.0.1
GitPython==3.1.44
pytest==8.4.0
pytest-asyncio==0.26.0